+++ include/ignition/fuel_tools/FuelClient.hh
@@ -73,6 +73,16 @@ namespace ignition
       public: Result ModelDetails(const ModelIdentifier &_id,
                                   ModelIdentifier &_model) const;
 
+      /// \brief Fetch the details of a model.
+      /// \param[in] _id a partially filled out identifier used to fetch models
+      /// \remarks Fulfills Get-One requirement
+      /// \param[out] _model The requested model
+      /// \return Result of the fetch operation.
+      public: Result ModelDetails(const ModelIdentifier &_id,
+                  ModelIdentifier &_model,
+                  const std::vector<std::string> &_headers) const;
+
+
       /// \brief Returns an iterator that can return names of models
       /// \remarks Fulfills Get-All requirement
       /// \remarks an iterator instead of a list of names is returned in case
@@ -114,7 +124,7 @@ namespace ignition
       /// \param[in] _id a partially filled out identifier used to fetch models
       /// \remarks Fulfills Get-One requirement
       /// \remarks It's not yet clear if model names are unique, so this API
-      ///          allows the posibility of getting multiple models with the
+      ///          allows the possibility of getting multiple models with the
       ///          same name.
       /// \return An iterator of models with names matching the criteria
       public: ModelIter Models(const ModelIdentifier &_id);
@@ -123,7 +133,7 @@ namespace ignition
       /// \param[in] _id a partially filled out identifier used to fetch models
       /// \remarks Fulfills Get-One requirement
       /// \remarks It's not yet clear if model names are unique, so this API
-      ///          allows the posibility of getting multiple models with the
+      ///          allows the possibility of getting multiple models with the
       ///          same name.
       /// \return An iterator of models with names matching the criteria
       public: ModelIter Models(const ModelIdentifier &_id) const;
@@ -191,7 +201,7 @@ namespace ignition
       /// \brief Download a world from ignition fuel. This will override an
       /// existing local copy of the world.
       /// \param[in] _worldUrl The unique URL of the world to download.
-      /// E.g.: https://fuel.ignitionrobotics.org/1.0/openrobotics/worlds/Empty
+      /// E.g.: https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Empty
       /// \param[out] _path Path where the world was downloaded.
       /// \return Result of the download operation.
       public: Result DownloadWorld(const common::URI &_worldUrl,
@@ -213,7 +223,7 @@ namespace ignition
 
       /// \brief Check if a world is already present in the local cache.
       /// \param[in] _worldUrl The unique URL of the world on a Fuel server.
-      /// E.g.: https://fuel.ignitionrobotics.org/1.0/openrobotics/worlds/Empty
+      /// E.g.: https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Empty
       /// \param[out] _path Local path where the world can be found.
       /// \return FETCH_ERROR if not cached, FETCH_ALREADY_EXISTS if cached.
       public: Result CachedWorld(const common::URI &_worldUrl,
@@ -221,7 +231,7 @@ namespace ignition
 
       /// \brief Check if a world exists in the cache.
       /// \param[in] _worldUrl The unique URL of the world on a Fuel server.
-      /// E.g.: https://fuel.ignitionrobotics.org/1.0/openrobotics/worlds/Empty
+      /// E.g.: https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Empty
       /// \return True if the world exists in the cache, false otherwise.
       public: bool CachedWorld(const common::URI &_worldUrl);
 
@@ -243,7 +253,7 @@ namespace ignition
       public: Result CachedWorldFile(const common::URI &_fileUrl,
                                      std::string &_path);
 
-      /// \brief Parse model identifer from model URL or unique name.
+      /// \brief Parse model identifier from model URL or unique name.
       /// \param[in] _modelUrl The unique URL of a model. It may also be a
       /// unique name, which is a URL without the server version.
       /// \param[out] _id The model identifier. It may contain incomplete
@@ -255,7 +265,7 @@ namespace ignition
       public: bool ParseModelUrl(const common::URI &_modelUrl,
                                  ModelIdentifier &_id);
 
-      /// \brief Parse world identifer from world URL or unique name.
+      /// \brief Parse world identifier from world URL or unique name.
       /// \param[in] _worldUrl The unique URL of a world. It may also be a
       /// unique name, which is a URL without the server version.
       /// \param[out] _id The world identifier. It may contain incomplete
@@ -267,7 +277,7 @@ namespace ignition
       public: bool ParseWorldUrl(const common::URI &_worldUrl,
                                  WorldIdentifier &_id);
 
-      /// \brief Parse model file identifer from model file URL.
+      /// \brief Parse model file identifier from model file URL.
       /// \param[in] _modelFileUrl The unique URL of a model file. It may also
       /// be a unique name, which is a URL without the server version.
       /// \param[out] _id The model identifier. It may contain incomplete
@@ -280,7 +290,7 @@ namespace ignition
                                      ModelIdentifier &_id,
                                      std::string &_filePath);
 
-      /// \brief Parse world file identifer from world file URL.
+      /// \brief Parse world file identifier from world file URL.
       /// \param[in] _worldFileUrl The unique URL of a world file. It may also
       /// be a unique name, which is a URL without the server version.
       /// \param[out] _id The world identifier. It may contain incomplete
@@ -292,6 +302,31 @@ namespace ignition
       public: bool ParseWorldFileUrl(const common::URI &_worldFileUrl,
                                      WorldIdentifier &_id,
                                      std::string &_filePath);
+
+      /// \brief This function requests the available licenses from the
+      ///  Fuel server and stores this information locally.
+      ///
+      /// The UploadModel function can use this information to set
+      /// appropriate license information based on a model's metadata.pbtxt
+      /// file. If license information is not available, then the
+      /// UploadModel function will default to the
+      /// "Creative Commons - Public Domain" license.
+      /// \param[in] _server Information about the server that provides
+      /// license information.
+      public: void PopulateLicenses(const ServerConfig &_server);
+
+      /// \brief Update a model using a PATCH request.
+      ///
+      /// Model fields that are patched by this function:
+      ///   * private
+      ///
+      /// \param[in] _model The model to patch. The contents of this model
+      ///            will be sent in the PATCH request.
+      /// \param[in] _headers Headers to set on the HTTP request.
+      /// \return Result of the patch operation.
+      public: Result PatchModel(
+                  const ignition::fuel_tools::ModelIdentifier &_model,
+                  const std::vector<std::string> &_headers);
 
       /// \brief PIMPL
       private: std::unique_ptr<FuelClientPrivate> dataPtr;