+++ src/FuelClient_TEST.cc
@@ -287,13 +287,13 @@ TEST_F(FuelClientTest, ParseModelFileURL)
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://fuel.ignitionrobotics.org/1.0/openrobotics/models/"
+      "https://fuel.ignitionrobotics.org/1.0/OpenRobotics/models/"
         "Cordless Drill/tip/files/meshes/cordless_drill.dae"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Cordless Drill");
     EXPECT_EQ(filePath, "meshes/cordless_drill.dae");
   }
@@ -306,13 +306,13 @@ TEST_F(FuelClientTest, ParseModelFileURL)
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://fuel.ignitionrobotics.org/1.0/openrobotics/models/Pine Tree/tip/"
+      "https://fuel.ignitionrobotics.org/1.0/OpenRobotics/models/Pine Tree/tip/"
       "files/materials/scripts/pine_tree.material"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Pine Tree");
     EXPECT_EQ(filePath, "materials/scripts/pine_tree.material");
   }
@@ -325,13 +325,13 @@ TEST_F(FuelClientTest, ParseModelFileURL)
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://fuel.ignitionrobotics.org/5.0/openrobotics/models/Pine Tree/tip/"
+      "https://fuel.ignitionrobotics.org/5.0/OpenRobotics/models/Pine Tree/tip/"
       "files/model.sdf"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Pine Tree");
     EXPECT_EQ(filePath, "model.sdf");
   }
@@ -342,14 +342,14 @@ TEST_F(FuelClientTest, ParseModelFileURL)
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://fuel.ignitionrobotics.org/openrobotics/models/Pine Tree/tip/"
+      "https://fuel.ignitionrobotics.org/OpenRobotics/models/Pine Tree/tip/"
       "files/materials/scripts/pine_tree.material"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_FALSE(id.Server().Version().empty());
     EXPECT_EQ("1.0", id.Server().Version());
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Pine Tree");
     EXPECT_EQ(filePath, "materials/scripts/pine_tree.material");
   }
@@ -362,13 +362,13 @@ TEST_F(FuelClientTest, ParseModelFileURL)
     ModelIdentifier id;
     std::string filePath;
     const common::URI modelUrl{
-      "https://fuel.ignitionrobotics.org/openrobotics/models/Pine Tree/tip/"
+      "https://fuel.ignitionrobotics.org/OpenRobotics/models/Pine Tree/tip/"
       "files/materials/scripts/pine_tree.material"};
     EXPECT_TRUE(client.ParseModelFileUrl(modelUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Pine Tree");
     EXPECT_EQ(filePath, "materials/scripts/pine_tree.material");
   }
@@ -809,15 +809,15 @@ TEST_F(FuelClientTest, ParseWorldFileUrl)
     WorldIdentifier id;
     std::string filePath;
     const common::URI worldUrl{
-      "https://fuel.ignitionrobotics.org/1.0/openrobotics/worlds/Empty/tip/"
-      "files/empty.world"};
+      "https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Empty/tip/"
+      "files/test.world"};
     EXPECT_TRUE(client.ParseWorldFileUrl(worldUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Empty");
-    EXPECT_EQ(filePath, "empty.world");
+    EXPECT_EQ(filePath, "test.world");
   }
 
   // URL - with client config
@@ -828,13 +828,13 @@ TEST_F(FuelClientTest, ParseWorldFileUrl)
     WorldIdentifier id;
     std::string filePath;
     const common::URI worldUrl{
-      "https://fuel.ignitionrobotics.org/1.0/openrobotics/worlds/Empty sky/tip/"
+      "https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Empty sky/tip/"
       "files/empty_sky.world"};
     EXPECT_TRUE(client.ParseWorldFileUrl(worldUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Empty sky");
     EXPECT_EQ(filePath, "empty_sky.world");
   }
@@ -847,15 +847,15 @@ TEST_F(FuelClientTest, ParseWorldFileUrl)
     WorldIdentifier id;
     std::string filePath;
     const common::URI worldUrl{
-      "https://fuel.ignitionrobotics.org/5.0/openrobotics/worlds/Empty/tip/"
-      "files/empty.world"};
+      "https://fuel.ignitionrobotics.org/5.0/OpenRobotics/worlds/Empty/tip/"
+      "files/test.world"};
     EXPECT_TRUE(client.ParseWorldFileUrl(worldUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Empty");
-    EXPECT_EQ(filePath, "empty.world");
+    EXPECT_EQ(filePath, "test.world");
   }
 
   // Unique name - without client config
@@ -864,14 +864,14 @@ TEST_F(FuelClientTest, ParseWorldFileUrl)
     WorldIdentifier id;
     std::string filePath;
     const common::URI worldUrl{
-      "https://fuel.ignitionrobotics.org/openrobotics/worlds/Empty sky/tip/"
+      "https://fuel.ignitionrobotics.org/OpenRobotics/worlds/Empty sky/tip/"
       "files/empty_sky.world"};
     EXPECT_TRUE(client.ParseWorldFileUrl(worldUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_FALSE(id.Server().Version().empty());
     EXPECT_EQ("1.0", id.Server().Version());
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Empty sky");
     EXPECT_EQ(filePath, "empty_sky.world");
   }
@@ -884,15 +884,15 @@ TEST_F(FuelClientTest, ParseWorldFileUrl)
     WorldIdentifier id;
     std::string filePath;
     const common::URI worldUrl{
-      "https://fuel.ignitionrobotics.org/1.0/openrobotics/worlds/Empty/tip/"
-      "files/empty.world"};
+      "https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Empty/tip/"
+      "files/test.world"};
     EXPECT_TRUE(client.ParseWorldFileUrl(worldUrl, id, filePath));
 
     EXPECT_EQ(id.Server().Url().Str(), "https://fuel.ignitionrobotics.org");
     EXPECT_EQ(id.Server().Version(), "1.0");
-    EXPECT_EQ(id.Owner(), "openrobotics");
+    EXPECT_EQ(id.Owner(), "OpenRobotics");
     EXPECT_EQ(id.Name(), "Empty");
-    EXPECT_EQ(filePath, "empty.world");
+    EXPECT_EQ(filePath, "test.world");
   }
 
   // Bad URL
@@ -924,7 +924,7 @@ TEST_F(FuelClientTest, DownloadWorld)
 
   ServerConfig server;
   server.SetUrl(ignition::common::URI(
-        "https://staging-fuel.ignitionrobotics.org"));
+        "https://fuel.ignitionrobotics.org"));
 
   ClientConfig config;
   config.AddServer(server);
@@ -937,8 +937,8 @@ TEST_F(FuelClientTest, DownloadWorld)
   // Download world from URL
   {
     // Unversioned URL should get the latest available version
-    common::URI url{
-        "https://staging-fuel.ignitionrobotics.org/1.0/nate/worlds/Empty"};
+    common::URI url{"https://fuel.ignitionrobotics.org/1.0/OpenRobotics/"
+                    "worlds/Test world"};
 
     // Check it is not cached
     std::string cachedPath;
@@ -954,31 +954,32 @@ TEST_F(FuelClientTest, DownloadWorld)
 
     // Check it was downloaded to `1`
     EXPECT_EQ(path, common::cwd() +
-        "/test_cache/staging-fuel.ignitionrobotics.org/nate/worlds/Empty/1");
+        "/test_cache/fuel.ignitionrobotics.org/OpenRobotics/worlds/"
+        "Test world/1");
     EXPECT_TRUE(common::exists(
-        "test_cache/staging-fuel.ignitionrobotics.org/nate/worlds/Empty/1"));
+        "test_cache/fuel.ignitionrobotics.org/OpenRobotics/worlds/"
+        "Test world/1"));
     EXPECT_TRUE(common::exists(
-       "test_cache/staging-fuel.ignitionrobotics.org/nate/worlds/Empty/1/"
-       "empty.world"));
+       "test_cache/fuel.ignitionrobotics.org/OpenRobotics/worlds/"
+       "Test world/1/test.world"));
 
     // Check it wasn't downloaded to world root directory
     EXPECT_FALSE(common::exists(
-          "test_cache/staging-fuel.ignitionrobotics.org/" +
-          std::string("nate/worlds/Empty/empty.world")));
+          "test_cache/fuel.ignitionrobotics.org/" +
+          std::string("OpenRobotics/worlds/Test world/test.world")));
 
     // Check it is cached
     auto res3 = client.CachedWorld(url, cachedPath);
     EXPECT_TRUE(res3);
     EXPECT_EQ(Result(ResultType::FETCH_ALREADY_EXISTS), res3);
-    EXPECT_EQ(common::cwd() +
-      "/test_cache/staging-fuel.ignitionrobotics.org/nate/worlds/Empty/1",
-      cachedPath);
+    EXPECT_EQ(common::cwd() + "/test_cache/fuel.ignitionrobotics.org"
+      "/OpenRobotics/worlds/Test world/1", cachedPath);
   }
 
   // Try using nonexistent URL
   {
     common::URI url{
-        "https://staging-fuel.ignitionrobotics.org/1.0/nate/worlds/Bad world"};
+        "https://fuel.ignitionrobotics.org/1.0/OpenRobotics/worlds/Bad world"};
     std::string path;
     auto result = client.DownloadWorld(url, path);
     EXPECT_FALSE(result);